requirements.txt
PySide6
distro
psutil
pyuac
py-cpuinfo
pyqtgraph